requirement.txt
flask==3.0.3
werkzeug==3.0.4
opencv-python==4.9.0.80
deepface==0.0.93
mtcnn==1.0.0
python-dotenv==1.0.1
numpy==1.26.4
h5py==3.12.1
scipy==1.14.1
PyMySQL==1.1.1
Flask-JWT-Extended==4.7.1
PyJWT==2.10.0
tensorflow==2.18.0
keras==3.6.0
pandas==2.2.3